requirements.txt
Pyqt5==5.15.11
mysql-connector-python==9.0.0